feat(gulpfile): add gulp support for compiling browser js
move browser js functions into webfunctions.js. compile main.js and webfunctions.js into bundle.js
via browserify and gulp. add jquery eventlistener to adminWarning and change users/manage view

no
