+++ themes/codeforpoland/views/layout.jade
@@ -24,7 +24,7 @@ html
     script(src='https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.11.1/moment.min.js')
     script(src='https://cdnjs.cloudflare.com/ajax/libs/fullcalendar/2.6.0/fullcalendar.js')
     script(src='/js/lib/simplemde.min.js')
-    script(src='/js/main.js')
+    script(src='/js/bundle.js')
 
     // Google Analytics: change UA-XXXXX-X to be your site's ID
     script.
